move migration utils files
