+++ packages/data-context/src/sources/migration/index.ts
@@ -2,5 +2,7 @@
 // created by autobarrel, do not modify directly
 
 export * from './autoRename'
+export * from './format'
 export * from './regexps'
 export * from './shouldShowSteps'
+export * from './utils'
