+++ packages/data-context/src/util/index.ts
@@ -6,6 +6,4 @@ export * from './cached'
 export * from './config-file-updater'
 export * from './config-options'
 export * from './file'
-export * from './migration'
-export * from './migrationFormat'
 export * from './urqlCacheKeys'
